Base2cam fixes (#118)
This fixes a crash if there is a tag detection which is not in the rig. Also add command line parameters for setting the wheel radius and wheel baseline, and viewing direction, as some of our cameras are right side up versus upside down. This now converges on tractor0002 and tractor0003 data.